Move from GetFilteredCourses from controller to service
Pure refactor.
